+++ src/components/services/AccountCallbackManager.tsx
@@ -8,13 +8,15 @@ import { checkFioObtData, refreshConnectedWallets } from '../../actions/FioActio
 import { refreshAllFioAddresses } from '../../actions/FioAddressActions'
 import { showReceiveDropdown } from '../../actions/ReceiveDropdown'
 import { checkPasswordRecovery } from '../../actions/RecoveryReminderActions'
+import { updateDepositAmount, updateTransactionCount } from '../../actions/RequestReviewActions'
 import { useAsyncEffect } from '../../hooks/useAsyncEffect'
 import { useWalletsSubscriber } from '../../hooks/useWalletsSubscriber'
 import { stakeMetadataCache } from '../../plugins/stake-plugins/metadataCache'
-import { useDispatch } from '../../types/reactRedux'
+import { getExchangeDenomByCurrencyCode } from '../../selectors/DenominationSelectors'
+import { useDispatch, useSelector } from '../../types/reactRedux'
 import { NavigationBase } from '../../types/routerTypes'
 import { makePeriodicTask } from '../../util/PeriodicTask'
-import { datelog, snooze } from '../../util/utils'
+import { convertCurrencyFromExchangeRates, convertNativeToExchange, datelog, snooze } from '../../util/utils'
 import { Airship } from './AirshipInstance'
 
 const REFRESH_RATES_MS = 30000
@@ -38,6 +40,7 @@ const notDirty: DirtyList = {
 export function AccountCallbackManager(props: Props) {
   const { account, navigation } = props
   const dispatch = useDispatch()
+  const exchangeRates = useSelector(state => state.exchangeRates)
   const [dirty, setDirty] = React.useState<DirtyList>(notDirty)
   const numWallets = React.useRef(0)
 
@@ -107,6 +110,23 @@ export function AccountCallbackManager(props: Props) {
         // Check for incoming FIO requests:
         const receivedTxs = transactions.filter(tx => !tx.isSend)
         if (receivedTxs.length > 0) dispatch(checkFioObtData(wallet, receivedTxs)).catch(err => console.warn(err))
+
+        // Review triggers: deposit & transaction count
+        for (const tx of transactions) {
+          const actionType = tx.savedAction?.actionType ?? tx.chainAction?.actionType
+
+          if (!tx.isSend) {
+            dispatch(updateTransactionCount()).catch(err => console.warn(err))
+            const exchangeDenom = getExchangeDenomByCurrencyCode(wallet.currencyConfig, tx.currencyCode)
+            const cryptoAmount = Math.abs(parseFloat(convertNativeToExchange(exchangeDenom.multiplier)(tx.nativeAmount)))
+            const usdAmount = parseFloat(convertCurrencyFromExchangeRates(exchangeRates, tx.currencyCode, 'iso:USD', String(cryptoAmount)))
+            if (usdAmount > 0) {
+              dispatch(updateDepositAmount(usdAmount)).catch(err => console.warn(err))
+            }
+          } else if (actionType !== 'swap' && actionType !== 'fiat') {
+            dispatch(updateTransactionCount()).catch(err => console.warn(err))
+          }
+        }
 
         // Show the dropdown for the first transaction:
         const [firstReceive] = receivedTxs